+++ src/server/adapters/LocalDateTimeAdapter.java
@@ -1,0 +1,23 @@
+package server.adapters;
+
+import com.google.gson.TypeAdapter;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonWriter;
+
+import java.io.IOException;
+import java.time.LocalDateTime;
+
+import static tasks.Task.FORMATTER;
+
+public class LocalDateTimeAdapter extends TypeAdapter<LocalDateTime> {
+
+    @Override
+    public void write(final JsonWriter jsonWriter, final LocalDateTime localDate) throws IOException {
+        jsonWriter.value(localDate.format(FORMATTER));
+    }
+
+    @Override
+    public LocalDateTime read(final JsonReader jsonReader) throws IOException {
+        return LocalDateTime.parse(jsonReader.nextString(), FORMATTER);
+    }
+}
